split oversized regression suites
