+++ packages/react-doctor/tests/regressions/rn-package-scoping.test.ts
@@ -776,714 +776,3 @@ export const Screen = () => {
     expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
   });
 });
-
-describe("classify-package-platform: dependency-section coverage", () => {
-  it("classifies a package with `react-native` only in `peerDependencies` as react-native", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-peer-only", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0" },
-        peerDependencies: { "react-native": ">=0.74.0" },
-      },
-      files: {
-        "src/Screen.tsx": `import { View } from "react-native";\nexport const Screen = () => <View>raw</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("classifies a package with `react-native` only in `devDependencies` as react-native", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-dev-only", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0" },
-        devDependencies: { "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Screen.tsx": `import { View } from "react-native";\nexport const Screen = () => <View>raw</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("classifies a package with `react-native` only in `optionalDependencies` as react-native", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-optional-only", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0" },
-        optionalDependencies: { "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Screen.tsx": `import { View } from "react-native";\nexport const Screen = () => <View>raw</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("classifies a package with `expo-router` only as react-native (Expo Router implies an Expo/RN target)", async () => {
-    const projectDir = setupReactProject(tempRoot, "expo-router-only", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "expo-router": "^3.5.0" },
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>raw</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("treats a manifest with BOTH `react-native` AND `next` as react-native (RN priority — react-native-web shape)", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-and-next-mixed", {
-      packageJsonExtras: {
-        dependencies: {
-          react: "^19.0.0",
-          "react-native": "0.76.0",
-          next: "^14.0.0",
-          "react-dom": "^19.0.0",
-        },
-      },
-      files: {
-        "src/Screen.tsx": `import { View } from "react-native";\nexport const Screen = () => <View>raw on RN</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    // RN wins over Next when both are declared — the package targets
-    // mobile too, so rules must keep firing.
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-});
-
-describe("classify-package-platform: malformed and empty manifest fallbacks", () => {
-  it("treats a malformed package.json as 'unknown' and falls back to the project-level framework hint (active)", async () => {
-    const projectDir = path.join(tempRoot, "malformed-pkg");
-    fs.mkdirSync(path.join(projectDir, "src"), { recursive: true });
-    fs.writeFileSync(path.join(projectDir, "package.json"), "{ not valid json,,,");
-    writeFile(
-      path.join(projectDir, "tsconfig.json"),
-      JSON.stringify({
-        compilerOptions: { jsx: "preserve", target: "es2022", module: "esnext" },
-      }),
-    );
-    writeFile(
-      path.join(projectDir, "src", "Screen.tsx"),
-      `export const Screen = () => <View>raw</View>;\n`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("treats an empty `{}` package.json as 'unknown' and falls back to the project-level framework", async () => {
-    const projectDir = path.join(tempRoot, "empty-pkg");
-    fs.mkdirSync(path.join(projectDir, "src"), { recursive: true });
-    writeJson(path.join(projectDir, "package.json"), {});
-    writeJson(path.join(projectDir, "tsconfig.json"), {
-      compilerOptions: { jsx: "preserve", target: "es2022", module: "esnext" },
-    });
-    writeFile(
-      path.join(projectDir, "src", "Screen.tsx"),
-      `export const Screen = () => <View>raw</View>;\n`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("with the project framework set to 'nextjs' and a malformed sub-package.json, RN rules stay inactive (capability gate never enables them)", async () => {
-    const projectDir = path.join(tempRoot, "framework-next-malformed-pkg");
-    fs.mkdirSync(path.join(projectDir, "src"), { recursive: true });
-    fs.writeFileSync(path.join(projectDir, "package.json"), "{ bad");
-    writeJson(path.join(projectDir, "tsconfig.json"), {
-      compilerOptions: { jsx: "preserve", target: "es2022", module: "esnext" },
-    });
-    writeFile(
-      path.join(projectDir, "src", "Page.tsx"),
-      `export const Page = () => <View>raw</View>;\n`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "nextjs" }),
-    });
-    // The project-level capability gate (`requires: ["react-native"]`)
-    // prevents the rule from even loading when the project framework
-    // is `nextjs`. Pin that this remains true regardless of file-level
-    // ambiguity.
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-});
-
-describe("classify-package-platform: file-extension overrides win over package classification", () => {
-  it("`.web.tsx` is skipped even when the enclosing package declares `react-native`", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-pkg-with-web-extension", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Screen.web.tsx": `export const Screen = () => <View>web</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-
-  it("`.ios.tsx` is scanned even when the enclosing package declares a web framework (force-on for native targets)", async () => {
-    const projectDir = setupReactProject(tempRoot, "web-pkg-with-ios-extension", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-dom": "^19.0.0", next: "^14.0.0" },
-      },
-      files: {
-        "src/Screen.ios.tsx": `export const Screen = () => <View>iOS</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("`.web.jsx` is skipped (matches the same `.web.[cm]?[jt]sx?` pattern as `.web.tsx`)", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-pkg-web-jsx-extension", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Screen.web.jsx": `export const Screen = () => <View>jsx web</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-
-  it("`.web.tsx.bak` or other suffixed names do NOT trigger the web-extension skip (regex anchored to end-of-name)", async () => {
-    // The .bak suffix means oxlint won't even scan the file (not a
-    // source extension), so we instead use a filename that DOES end
-    // in .tsx but contains `.web` as a non-trailing segment.
-    const projectDir = setupReactProject(tempRoot, "rn-pkg-confusable-name", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        // `Screen.web.theme.tsx` does NOT end with `.web.tsx`, so the
-        // web-extension shortcut should NOT trigger — the file should
-        // be scanned as a normal RN file.
-        "src/Screen.web.theme.tsx": `export const Screen = () => <View>not web-only</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-});
-
-describe("project-level framework fallback when package classification is ambiguous", () => {
-  it("with `framework: 'expo'` at the project level and an ambiguous package, RN rules fire", async () => {
-    const projectDir = setupReactProject(tempRoot, "ambiguous-pkg-expo-fallback", {
-      packageJsonExtras: {
-        // No RN or web framework keys, just plain react.
-        dependencies: { react: "^19.0.0" },
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>raw</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "expo" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-});
-
-describe("classify-package-platform: cache safety across independent packages", () => {
-  it("classifies two same-cased package names in distinct directories independently (cache key is the directory path)", async () => {
-    const rnDir = setupReactProject(tempRoot, "cache-safety-rn-A", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>RN raw</View>;\n`,
-      },
-    });
-    const webDir = setupReactProject(tempRoot, "cache-safety-web-A", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-dom": "^19.0.0", next: "^14.0.0" },
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>web</View>;\n`,
-      },
-    });
-
-    const rnDiagnostics = await runOxlint({
-      rootDirectory: rnDir,
-      project: buildTestProject({ rootDirectory: rnDir, framework: "react-native" }),
-    });
-    const webDiagnostics = await runOxlint({
-      rootDirectory: webDir,
-      project: buildTestProject({ rootDirectory: webDir, framework: "react-native" }),
-    });
-
-    expect(findDiagnosticsByRule(rnDiagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-    expect(findDiagnosticsByRule(webDiagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-});
-
-describe('rn-no-raw-text: "use dom" directive composition with package gating', () => {
-  it("`use dom` short-circuits even inside an RN package", async () => {
-    const projectDir = setupReactProject(tempRoot, "use-dom-in-rn-pkg", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        "src/Dom.tsx": `"use dom";\nexport const Dom = () => <div>Web rendered</div>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-});
-
-describe("rn-* rules other than rn-no-raw-text: package gating", () => {
-  it("rn-prefer-pressable is silenced on a Next.js sub-package that legitimately imports from react-native (e.g. via solito/react-native-web)", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-prefer-pressable-on-web", {
-      packageJsonExtras: {
-        dependencies: {
-          react: "^19.0.0",
-          "react-dom": "^19.0.0",
-          next: "^14.0.0",
-        },
-      },
-      files: {
-        // TouchableOpacity import would normally trip rn-prefer-pressable.
-        "src/Web.tsx": `import { TouchableOpacity } from "react-native";\nexport const Web = () => <TouchableOpacity />;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-prefer-pressable")).toHaveLength(0);
-  });
-
-  it("rn-no-deprecated-modules fires on an RN package even when sibling web sub-packages exist", async () => {
-    const projectDir = setupReactProject(tempRoot, "rn-deprecated-mobile-with-web-sibling", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0", "react-native": "0.76.0" },
-      },
-      files: {
-        "src/index.ts": `import { AsyncStorage } from "react-native";\nvoid AsyncStorage;\n`,
-      },
-    });
-    // Drop a web-only sibling package nearby.
-    writeJson(path.join(projectDir, "siblings", "web-sibling", "package.json"), {
-      name: "web-sibling",
-      dependencies: { next: "^14.0.0", react: "^19.0.0", "react-dom": "^19.0.0" },
-    });
-    writeFile(
-      path.join(projectDir, "siblings", "web-sibling", "src", "Web.tsx"),
-      `export const Web = () => <View>web</View>;\n`,
-    );
-
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    const deprecatedHits = findDiagnosticsByRule(diagnostics, "rn-no-deprecated-modules");
-    expect(deprecatedHits.length).toBeGreaterThan(0);
-    // No rn-* diagnostic should be reported against the web sibling.
-    const webSiblingRn = diagnostics.filter(
-      (diagnostic) =>
-        diagnostic.rule.startsWith("rn-") &&
-        diagnostic.filePath.replaceAll("\\", "/").includes("siblings/web-sibling/"),
-    );
-    expect(webSiblingRn).toHaveLength(0);
-  });
-});
-
-describe("React Native package classification: namespace prefixes and Metro `react-native` field", () => {
-  it("classifies a package as RN when it declares a `@react-native-firebase/*` dependency (prefix match)", async () => {
-    // `@react-native-firebase/app` lives under the `@react-native-`
-    // community namespace. The closed-set check in the previous
-    // implementation missed it; the prefix match recognises every
-    // member of the namespace without us having to enumerate them.
-    const projectDir = setupReactProject(tempRoot, "rn-namespace-prefix-match", {
-      packageJsonExtras: {
-        dependencies: {
-          react: "^19.0.0",
-          "@react-native-firebase/app": "^21.0.0",
-        },
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>RN firebase package</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it("classifies a package as RN when it declares Metro's top-level `react-native` resolution field (library manifest)", async () => {
-    // RN-only libraries set the top-level `react-native` field so Metro
-    // resolves an RN-specific entry over `main` / `module`. The mere
-    // presence of the field is enough to classify the package as
-    // RN-aware even when its dep list is otherwise empty.
-    const projectDir = setupReactProject(tempRoot, "rn-metro-resolution-field", {
-      packageJsonExtras: {
-        dependencies: { react: "^19.0.0" },
-        "react-native": "./dist/native/index.js",
-      },
-      files: {
-        "src/Screen.tsx": `export const Screen = () => <View>Metro field library</View>;\n`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-});
-
-describe("rn-no-raw-text: Platform.select fork (canonical RN platform-fork helper)", () => {
-  const setupPlatformSelectProject = (caseId: string, sourceCode: string): string =>
-    setupReactProject(tempRoot, caseId, {
-      packageJsonExtras: { dependencies: { react: "^19.0.0", "react-native": "0.76.0" } },
-      files: { "src/Screen.tsx": sourceCode },
-    });
-
-  it("skips raw text in the `web` arm of `Platform.select({ web: ..., default: ... })`", async () => {
-    const projectDir = setupPlatformSelectProject(
-      "platform-select-web-arm",
-      `import { Platform, View } from "react-native";
-
-export const Screen = () =>
-  Platform.select({
-    web: <View>Web fallback markup</View>,
-    default: null,
-  });
-`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-
-  it("STILL fires raw text in non-web arms of `Platform.select` (`default`, `ios`, `android`)", async () => {
-    const projectDir = setupPlatformSelectProject(
-      "platform-select-default-arm-still-fires",
-      `import { Platform, View } from "react-native";
-
-export const Screen = () =>
-  Platform.select({
-    web: null,
-    default: <View>Native fallback that crashes</View>,
-  });
-`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-
-  it('recognises the string-quoted property key form `{ "web": ... }`', async () => {
-    const projectDir = setupPlatformSelectProject(
-      "platform-select-string-key",
-      `import { Platform, View } from "react-native";
-
-export const Screen = () =>
-  Platform.select({
-    "web": <View>Web</View>,
-    default: null,
-  });
-`,
-    );
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-});
-
-describe("rn-no-raw-text: TS / chain wrapping around the Platform.OS read", () => {
-  it('recognises `Platform?.OS === "web"` (optional chain)', async () => {
-    const projectDir = setupReactProject(tempRoot, "platform-os-optional-chain", {
-      packageJsonExtras: { dependencies: { react: "^19.0.0", "react-native": "0.76.0" } },
-      files: {
-        "src/Screen.tsx": `import { Platform, View } from "react-native";
-
-export const Screen = () => {
-  if (Platform?.OS === "web") {
-    return <View>Optional-chain web</View>;
-  }
-  return null;
-};
-`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-
-  it('recognises `Platform.OS! === "web"` (TS non-null assertion)', async () => {
-    const projectDir = setupReactProject(tempRoot, "platform-os-non-null-assertion", {
-      packageJsonExtras: { dependencies: { react: "^19.0.0", "react-native": "0.76.0" } },
-      files: {
-        "src/Screen.tsx": `import { Platform, View } from "react-native";
-
-export const Screen = () => {
-  if (Platform.OS! === "web") {
-    return <View>Non-null asserted web</View>;
-  }
-  return null;
-};
-`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text")).toHaveLength(0);
-  });
-});
-
-describe("rn-no-raw-text: scope-boundary pruning (no nested-helper false negatives)", () => {
-  it("STILL fires raw text inside a callback hoisted out of a Platform.OS branch", async () => {
-    // `renderInner`'s body sits inside the `Platform.OS === "web"`
-    // alternate from the caller's perspective, but the function itself
-    // is hoisted to top level — it could be called from anywhere, so
-    // its raw text must not inherit the parent branch's exemption. The
-    // walker stops at the function boundary.
-    const projectDir = setupReactProject(tempRoot, "platform-os-scope-boundary-pruning", {
-      packageJsonExtras: { dependencies: { react: "^19.0.0", "react-native": "0.76.0" } },
-      files: {
-        "src/Screen.tsx": `import { Platform, View } from "react-native";
-
-const renderInner = () => <View>Hoisted raw text</View>;
-
-export const Screen = () => (Platform.OS === "web" ? renderInner() : null);
-`,
-      },
-    });
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: buildTestProject({ rootDirectory: projectDir, framework: "react-native" }),
-    });
-    expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-  });
-});
-
-describe("inverted monorepo: web-rooted project with an RN workspace still loads rn-* rules", () => {
-  it("fires rn-no-raw-text on the mobile workspace even when the entry-point `framework` is web-only", async () => {
-    // The entry-point project's framework resolves to `nextjs` here
-    // (we pass `framework: "nextjs"` explicitly), but
-    // `hasReactNativeWorkspace` is `true` because the discovered
-    // workspaces include an Expo app. Without the inverted-gate fix
-    // the capability builder dropped every `requires: ["react-native"]`
-    // rule before the file-level wrapper ever got to run.
-    const projectDir = setupReactProject(tempRoot, "inverted-monorepo-web-rooted", {
-      packageJsonExtras: {
-        dependencies: { next: "^14.0.0", react: "^19.0.0", "react-dom": "^19.0.0" },
-        // pnpm-shaped workspaces field also works; npm/yarn flat list
-        // is more portable across the test fixtures.
-        workspaces: ["apps/*"],
-      },
-      files: {
-        "apps/web/src/Page.tsx": `export const Page = () => <View>Web entry</View>;\n`,
-      },
-    });
-    writeJson(path.join(projectDir, "apps", "web", "package.json"), {
-      name: "web",
-      dependencies: { next: "^14.0.0", react: "^19.0.0", "react-dom": "^19.0.0" },
-    });
-    writeJson(path.join(projectDir, "apps", "mobile", "package.json"), {
-      name: "mobile",
-      dependencies: { react: "^19.0.0", "react-native": "0.76.0", expo: "^51.0.0" },
-    });
-    writeFile(
-      path.join(projectDir, "apps", "mobile", "src", "Screen.tsx"),
-      `import { View } from "react-native";\nexport const Screen = () => <View>Mobile entry</View>;\n`,
-    );
-
-    const diagnostics = await runOxlint({
-      rootDirectory: projectDir,
-      project: {
-        ...buildTestProject({ rootDirectory: projectDir, framework: "nextjs" }),
-        hasReactNativeWorkspace: true,
-      },
-    });
-
-    const rnHits = findDiagnosticsByRule(diagnostics, "rn-no-raw-text");
-    const normalizedPaths = rnHits.map((diagnostic) => diagnostic.filePath.replaceAll("\\", "/"));
-    expect(normalizedPaths.some((filePath) => filePath.includes("apps/mobile/"))).toBe(true);
-    expect(normalizedPaths.some((filePath) => filePath.includes("apps/web/"))).toBe(false);
-  });
-});
-
-// Each entry encodes a single RN "signal" the two predicates have to
-// agree on — added once here, exercised against BOTH layers below.
-//
-//   project-info side: `discoverProject(...).hasReactNativeWorkspace`
-//                      gates whether the `react-native` capability is
-//                      added in `buildCapabilities`, deciding if any
-//                      `rn-*` rule loads at all.
-//
-//   oxlint plugin side: `classifyPackagePlatform` (called by every
-//                       `rn-*` rule via `wrapReactNativeRule`)
-//                       decides whether a given file's nearest
-//                       `package.json` is treated as RN-aware.
-//
-// When someone adds a new package name to one side and forgets the
-// other, the sync table below catches the drift on its next test run.
-interface ReactNativeSignal {
-  description: string;
-  // `packageJsonExtras` is splatted into the workspace package.json.
-  // Use `dependencies` / `devDependencies` / `peerDependencies` /
-  // `optionalDependencies` / `"react-native"` to pin the RN signal
-  // shape under test.
-  packageJsonExtras: Partial<PackageJson> & { "react-native"?: string };
-}
-
-const REACT_NATIVE_SIGNALS: ReadonlyArray<ReactNativeSignal> = [
-  {
-    description: "bare `react-native` dependency",
-    packageJsonExtras: { dependencies: { "react-native": "0.76.0" } },
-  },
-  {
-    description: "`react-native-tvos` dependency",
-    packageJsonExtras: { dependencies: { "react-native-tvos": "0.76.0-0" } },
-  },
-  {
-    description: "`expo` + `expo-router` dependencies",
-    packageJsonExtras: {
-      dependencies: { expo: "^51.0.0", "expo-router": "^3.5.0" },
-    },
-  },
-  {
-    description: "`react-native` in `peerDependencies` only",
-    packageJsonExtras: { peerDependencies: { "react-native": "*" } },
-  },
-  {
-    description: "`react-native` in `devDependencies` only",
-    packageJsonExtras: { devDependencies: { "react-native": "0.76.0" } },
-  },
-  {
-    description: "`react-native` in `optionalDependencies` only",
-    packageJsonExtras: { optionalDependencies: { "react-native": "0.76.0" } },
-  },
-  {
-    description: "`@react-native-firebase/app` namespace dependency",
-    packageJsonExtras: { dependencies: { "@react-native-firebase/app": "^21.0.0" } },
-  },
-  {
-    description: "`@react-native/metro-config` namespace dependency",
-    packageJsonExtras: { devDependencies: { "@react-native/metro-config": "^0.76.0" } },
-  },
-  {
-    description: "Metro's top-level `react-native` resolution field (library manifest)",
-    packageJsonExtras: { "react-native": "./dist/native/index.js" },
-  },
-];
-
-describe("RN signal sync table: project-info and oxlint plugin RN detection MUST agree", () => {
-  // Project-info side — discoverProject is what populates
-  // `hasReactNativeWorkspace`. Exercising it inside a Next-rooted
-  // workspace catches signals that classify the file as RN
-  // (oxlint plugin) but never get the capability added
-  // (project-info) — the silent drop that prevents the rule from
-  // ever running.
-  for (const signal of REACT_NATIVE_SIGNALS) {
-    it(`discoverProject — ${signal.description} → hasReactNativeWorkspace: true`, () => {
-      const rootDirectory = fs.mkdtempSync(path.join(tempRoot, "rn-signal-discover-"));
-      const mobileDirectory = path.join(rootDirectory, "apps", "mobile");
-      fs.mkdirSync(mobileDirectory, { recursive: true });
-      writeJson(path.join(rootDirectory, "package.json"), {
-        name: "sync-root",
-        dependencies: { next: "^14.0.0", react: "^19.0.0", "react-dom": "^19.0.0" },
-        workspaces: ["apps/*"],
-      });
-      writeJson(path.join(mobileDirectory, "package.json"), {
-        name: "mobile",
-        dependencies: { react: "^19.0.0" },
-        ...signal.packageJsonExtras,
-      });
-
-      const projectInfo = discoverProject(rootDirectory);
-      expect(projectInfo.hasReactNativeWorkspace).toBe(true);
-    });
-  }
-
-  // Plugin side — classifyPackagePlatform is called inside the rule
-  // wrapper; the observable is whether `rn-no-raw-text` fires on a
-  // file inside the workspace. Exercising it here catches the
-  // inverse drift — a signal that's caught at the project level but
-  // missed at the file level, which would mean rn-* rules load
-  // globally and then over-fire on the same RN package.
-  for (const signal of REACT_NATIVE_SIGNALS) {
-    it(`classifyPackagePlatform — ${signal.description} → file classifies as RN`, async () => {
-      const projectDirectory = setupReactProject(
-        tempRoot,
-        `rn-signal-classify-${signal.description.replace(/[^a-z0-9]/gi, "-")}`,
-        {
-          packageJsonExtras: {
-            ...signal.packageJsonExtras,
-            dependencies: {
-              react: "^19.0.0",
-              ...(signal.packageJsonExtras.dependencies ?? {}),
-            },
-          },
-          files: {
-            "src/Screen.tsx": `export const Screen = () => <View>raw</View>;\n`,
-          },
-        },
-      );
-
-      const diagnostics = await runOxlint({
-        rootDirectory: projectDirectory,
-        project: buildTestProject({ rootDirectory: projectDirectory, framework: "react-native" }),
-      });
-      expect(findDiagnosticsByRule(diagnostics, "rn-no-raw-text").length).toBeGreaterThan(0);
-    });
-  }
-});
